feat: only keep polygon outlines that are required for connecting skeletons
With this change, we no longer need to include the entire polygon border. This is good since including the border is not necessarily desirable (e.g., road pavement edges are unneeded because there are likely already centerlines). We now keep only the broders required to connect skeleton lines. The skeleton lines cannot go outside the polygon, so any part of the skeleton line that is erased due to being outside the polygon is replaced with the polygon border for that segment.